refactor: 💡 csv parser
